Move content from constants/ to a new store/ directory
