--- src/constants/settings.js
@@ -1,1 +1,0 @@
-export const settings = chrome.storage.sync;
