+++ src/store/readSetting.js
@@ -1,0 +1,6 @@
+import { store } from 'store/store';
+import { defaultSettings } from 'store/defaultSettings';
+
+export default function readSetting(settingName, callback) {
+  return store.get({ [settingName]: defaultSettings[settingName] }, callback);
+}
