+++ src/store/store.js
@@ -1,0 +1,1 @@
+export const store = chrome.storage.sync;
