+++ src/store/writeSetting.js
@@ -1,0 +1,7 @@
+import { store } from 'store/store';
+import updateCheckbox from 'utils/updateCheckbox';
+
+export default function writeSetting(settingName, value) {
+  store.set({ [settingName]: value });
+  updateCheckbox(settingName, value);
+}
