+++ webpack.config.js
@@ -9,6 +9,7 @@ module.exports = {
       content: path.resolve(__dirname, 'src/content'),
       constants: path.resolve(__dirname, 'src/constants'),
       utils: path.resolve(__dirname, 'src/utils'),
+      store: path.resolve(__dirname, 'src/store'),
     },
   },
   entry: {
